refactor: Construction과 Sections의 강한 의존 제거
